requirements.txt
deeplake==3.4.3
einops==0.6.0
hydra-colorlog==1.2.0
hydra-core==1.2.0
hydra-optuna-sweeper==1.2.0
hydra-orion-sweeper==1.6.2
matplotlib==3.7.1
numpy==1.22.1
omegaconf==2.3.0
opt-einsum==3.3.0
optuna==2.10.1
orion==0.2.7
pytorch-lightning==1.8.6
scipy==1.10.1
seaborn==0.11.2
torch==1.13.1
torchdata==0.5.1
torchmetrics==0.11.0
torchtext==0.14.0
torchtyping==0.1.4
torchvision==0.14.0
tqdm==4.64.1
wandb==0.13.4